Refactor useEmulation and config-map access
