+++ pkg/virt-controller/services/template.go
@@ -23,26 +23,21 @@ import (
 	"fmt"
 	"path/filepath"
 	"strconv"
-	"strings"
 
 	k8sv1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/tools/cache"
+
+	"strings"
 
 	"kubevirt.io/kubevirt/pkg/api/v1"
 	"kubevirt.io/kubevirt/pkg/config"
+	"kubevirt.io/kubevirt/pkg/config-disk"
 	"kubevirt.io/kubevirt/pkg/hooks"
 	"kubevirt.io/kubevirt/pkg/precond"
 	"kubevirt.io/kubevirt/pkg/registry-disk"
 	"kubevirt.io/kubevirt/pkg/util/net/dns"
 )
-
-const configMapName = "kube-system/kubevirt-config"
-const UseEmulationKey = "debug.useEmulation"
-const ImagePullPolicyKey = "dev.imagePullPolicy"
-const KvmDevice = "devices.kubevirt.io/kvm"
-const TunDevice = "devices.kubevirt.io/tun"
 
 const CAP_NET_ADMIN = "NET_ADMIN"
 const CAP_SYS_NICE = "SYS_NICE"
@@ -55,46 +50,7 @@ type templateService struct {
 	launcherImage   string
 	virtShareDir    string
 	imagePullSecret string
-	store           cache.Store
-}
-
-func getConfigMapEntry(store cache.Store, key string) (string, error) {
-
-	if obj, exists, err := store.GetByKey(configMapName); err != nil {
-		return "", err
-	} else if !exists {
-		return "", nil
-	} else {
-		return obj.(*k8sv1.ConfigMap).Data[key], nil
-	}
-}
-
-func IsEmulationAllowed(store cache.Store) (useEmulation bool, err error) {
-	var value string
-	value, err = getConfigMapEntry(store, UseEmulationKey)
-	if strings.ToLower(value) == "true" {
-		useEmulation = true
-	}
-	return
-}
-
-func GetImagePullPolicy(store cache.Store) (policy k8sv1.PullPolicy, err error) {
-	var value string
-	if value, err = getConfigMapEntry(store, ImagePullPolicyKey); err != nil || value == "" {
-		policy = k8sv1.PullIfNotPresent // Default if not specified
-	} else {
-		switch value {
-		case "Always":
-			policy = k8sv1.PullAlways
-		case "Never":
-			policy = k8sv1.PullNever
-		case "IfNotPresent":
-			policy = k8sv1.PullIfNotPresent
-		default:
-			err = fmt.Errorf("Invalid ImagePullPolicy in ConfigMap: %s", value)
-		}
-	}
-	return
+	clusterConfig   *config.ClusterConfig
 }
 
 func (t *templateService) RenderLaunchManifest(vmi *v1.VirtualMachineInstance) (*k8sv1.Pod, error) {
@@ -195,7 +151,7 @@ func (t *templateService) RenderLaunchManifest(vmi *v1.VirtualMachineInstance) (
 			// attach a ConfigMap to the pod
 			volumesMounts = append(volumesMounts, k8sv1.VolumeMount{
 				Name:      volume.Name,
-				MountPath: filepath.Join(config.ConfigMapSourceDir, volume.Name),
+				MountPath: filepath.Join(config_disk.ConfigMapSourceDir, volume.Name),
 				ReadOnly:  true,
 			})
 			volumes = append(volumes, k8sv1.Volume{
@@ -213,7 +169,7 @@ func (t *templateService) RenderLaunchManifest(vmi *v1.VirtualMachineInstance) (
 			// attach a Secret to the pod
 			volumesMounts = append(volumesMounts, k8sv1.VolumeMount{
 				Name:      volume.Name,
-				MountPath: filepath.Join(config.SecretSourceDir, volume.Name),
+				MountPath: filepath.Join(config_disk.SecretSourceDir, volume.Name),
 				ReadOnly:  true,
 			})
 			volumes = append(volumes, k8sv1.Volume{
@@ -358,29 +314,13 @@ func (t *templateService) RenderLaunchManifest(vmi *v1.VirtualMachineInstance) (
 		"--hook-sidecars", strconv.Itoa(len(requestedHookSidecarList)),
 	}
 
-	useEmulation, err := IsEmulationAllowed(t.store)
-	if err != nil {
-		return nil, err
-	}
-
-	imagePullPolicy, err := GetImagePullPolicy(t.store)
+	imagePullPolicy, err := t.clusterConfig.GetImagePullPolicy()
 	if err != nil {
 		return nil, err
 	}
 
 	if resources.Limits == nil {
 		resources.Limits = make(k8sv1.ResourceList)
-	}
-
-	extraResources := getRequiredResources(vmi)
-	for key, val := range extraResources {
-		resources.Limits[key] = val
-	}
-
-	if useEmulation {
-		command = append(command, "--use-emulation")
-	} else {
-		resources.Limits[KvmDevice] = resource.MustParse("1")
 	}
 
 	// Add ports from interfaces to the pod manifest
@@ -553,14 +493,6 @@ func getRequiredCapabilities(vmi *v1.VirtualMachineInstance) []k8sv1.Capability 
 	return res
 }
 
-func getRequiredResources(vmi *v1.VirtualMachineInstance) k8sv1.ResourceList {
-	res := k8sv1.ResourceList{}
-	if (vmi.Spec.Domain.Devices.AutoattachPodInterface == nil) || (*vmi.Spec.Domain.Devices.AutoattachPodInterface == true) {
-		res[TunDevice] = resource.MustParse("1")
-	}
-	return res
-}
-
 func appendUniqueImagePullSecret(secrets []k8sv1.LocalObjectReference, newsecret k8sv1.LocalObjectReference) []k8sv1.LocalObjectReference {
 	for _, oldsecret := range secrets {
 		if oldsecret == newsecret {
@@ -646,13 +578,13 @@ func getMultusInterfaceList(vmi *v1.VirtualMachineInstance) string {
 	return strings.Join(ifaceList, ",")
 }
 
-func NewTemplateService(launcherImage string, virtShareDir string, imagePullSecret string, configMapCache cache.Store) TemplateService {
+func NewTemplateService(launcherImage string, virtShareDir string, imagePullSecret string, clusterConfig *config.ClusterConfig) TemplateService {
 	precond.MustNotBeEmpty(launcherImage)
 	svc := templateService{
 		launcherImage:   launcherImage,
 		virtShareDir:    virtShareDir,
 		imagePullSecret: imagePullSecret,
-		store:           configMapCache,
+		clusterConfig:   clusterConfig,
 	}
 	return &svc
 }
