+++ pkg/virt-controller/services/template_test.go
@@ -20,9 +20,6 @@
 package services_test
 
 import (
-	"testing"
-	"time"
-
 	. "kubevirt.io/kubevirt/pkg/virt-controller/services"
 
 	. "github.com/onsi/ginkgo"
@@ -31,11 +28,12 @@ import (
 	kubev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/watch"
 	"k8s.io/client-go/tools/cache"
 
+	"testing"
+
 	"kubevirt.io/kubevirt/pkg/api/v1"
+	"kubevirt.io/kubevirt/pkg/config"
 	"kubevirt.io/kubevirt/pkg/hooks"
 	"kubevirt.io/kubevirt/pkg/log"
 )
@@ -44,7 +42,7 @@ var _ = Describe("Template", func() {
 
 	log.Log.SetIOWriter(GinkgoWriter)
 	configCache := cache.NewIndexer(cache.DeletionHandlingMetaNamespaceKeyFunc, nil)
-	svc := NewTemplateService("kubevirt/virt-launcher", "/var/run/kubevirt", "pull-secret-1", configCache)
+	svc := NewTemplateService("kubevirt/virt-launcher", "/var/run/kubevirt", "pull-secret-1", config.NewClusterConfig(configCache))
 
 	Describe("Rendering", func() {
 		Context("launch template with correct parameters", func() {
@@ -679,86 +677,6 @@ var _ = Describe("Template", func() {
 			})
 		})
 
-		Context("with pod networking", func() {
-			It("Should require tun device by default", func() {
-				vmi := v1.VirtualMachineInstance{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "testvmi", Namespace: "default", UID: "1234",
-					},
-				}
-				pod, err := svc.RenderLaunchManifest(&vmi)
-				Expect(err).ToNot(HaveOccurred())
-
-				tun, ok := pod.Spec.Containers[0].Resources.Limits[TunDevice]
-				Expect(ok).To(BeTrue())
-				Expect(int(tun.Value())).To(Equal(1))
-
-				found := false
-				caps := pod.Spec.Containers[0].SecurityContext.Capabilities
-				for _, cap := range caps.Add {
-					if cap == CAP_NET_ADMIN {
-						found = true
-					}
-				}
-				Expect(found).To(BeTrue(), "Expected compute container to be granted NET_ADMIN capability")
-			})
-
-			It("Should require tun device if explicitly requested", func() {
-				domain := v1.DomainSpec{}
-				autoAttach := true
-				domain.Devices.AutoattachPodInterface = &autoAttach
-
-				vmi := v1.VirtualMachineInstance{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "testvmi", Namespace: "default", UID: "1234",
-					},
-					Spec: v1.VirtualMachineInstanceSpec{Domain: domain},
-				}
-				pod, err := svc.RenderLaunchManifest(&vmi)
-				Expect(err).ToNot(HaveOccurred())
-
-				tun, ok := pod.Spec.Containers[0].Resources.Limits[TunDevice]
-				Expect(ok).To(BeTrue())
-				Expect(int(tun.Value())).To(Equal(1))
-
-				found := false
-				caps := pod.Spec.Containers[0].SecurityContext.Capabilities
-				for _, cap := range caps.Add {
-					if cap == CAP_NET_ADMIN {
-						found = true
-					}
-				}
-				Expect(found).To(BeTrue(), "Expected compute container to be granted NET_ADMIN capability")
-			})
-
-			It("Should not require tun device if explicitly rejected", func() {
-				domain := v1.DomainSpec{}
-				autoAttach := false
-				domain.Devices.AutoattachPodInterface = &autoAttach
-
-				vmi := v1.VirtualMachineInstance{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "testvmi", Namespace: "default", UID: "1234",
-					},
-					Spec: v1.VirtualMachineInstanceSpec{Domain: domain},
-				}
-				pod, err := svc.RenderLaunchManifest(&vmi)
-				Expect(err).ToNot(HaveOccurred())
-
-				_, ok := pod.Spec.Containers[0].Resources.Limits[TunDevice]
-				Expect(ok).To(BeFalse())
-
-				found := false
-				caps := pod.Spec.Containers[0].SecurityContext.Capabilities
-				for _, cap := range caps.Add {
-					if cap == CAP_NET_ADMIN {
-						found = true
-					}
-				}
-				Expect(found).To(BeFalse(), "Expected compute container to not be granted NET_ADMIN capability")
-			})
-		})
-
 		Context("with a configMap volume source", func() {
 			It("Should add the ConfigMap to template", func() {
 				volumes := []v1.Volume{
@@ -819,143 +737,7 @@ var _ = Describe("Template", func() {
 			})
 		})
 	})
-	Describe("ConfigMap", func() {
-		var cmListWatch *cache.ListWatch
-		var cmInformer cache.SharedIndexInformer
-		var cmStore cache.Store
-		var stopChan chan struct{}
-
-		BeforeEach(func() {
-			stopChan = make(chan struct{})
-		})
-
-		AfterEach(func() {
-			close(stopChan)
-		})
-
-		It("Should return false if configmap is not present", func() {
-			cmListWatch = MakeFakeConfigMapWatcher([]kubev1.ConfigMap{})
-			cmInformer = cache.NewSharedIndexInformer(cmListWatch, &v1.VirtualMachineInstance{}, time.Second, cache.Indexers{cache.NamespaceIndex: cache.MetaNamespaceIndexFunc})
-			cmStore = cmInformer.GetStore()
-			go cmInformer.Run(stopChan)
-			cache.WaitForCacheSync(stopChan, cmInformer.HasSynced)
-
-			result, err := IsEmulationAllowed(cmStore)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(result).To(BeFalse())
-		})
-
-		It("Should return false if configmap doesn't have useEmulation set", func() {
-			cfgMap := kubev1.ConfigMap{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "kube-system",
-					Name:      "kubevirt-config",
-				},
-				Data: map[string]string{},
-			}
-			cmListWatch = MakeFakeConfigMapWatcher([]kubev1.ConfigMap{cfgMap})
-			cmInformer = cache.NewSharedIndexInformer(cmListWatch, &v1.VirtualMachineInstance{}, time.Second, cache.Indexers{cache.NamespaceIndex: cache.MetaNamespaceIndexFunc})
-			cmStore = cmInformer.GetStore()
-			go cmInformer.Run(stopChan)
-			cache.WaitForCacheSync(stopChan, cmInformer.HasSynced)
-
-			result, err := IsEmulationAllowed(cmStore)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(result).To(BeFalse())
-		})
-
-		It("Should return true if useEmulation = true", func() {
-			cfgMap := kubev1.ConfigMap{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "kube-system",
-					Name:      "kubevirt-config",
-				},
-				Data: map[string]string{UseEmulationKey: "true"},
-			}
-			cmListWatch = MakeFakeConfigMapWatcher([]kubev1.ConfigMap{cfgMap})
-			cmInformer = cache.NewSharedIndexInformer(cmListWatch, &v1.VirtualMachineInstance{}, time.Second, cache.Indexers{cache.NamespaceIndex: cache.MetaNamespaceIndexFunc})
-			cmStore = cmInformer.GetStore()
-			go cmInformer.Run(stopChan)
-			cache.WaitForCacheSync(stopChan, cmInformer.HasSynced)
-
-			result, err := IsEmulationAllowed(cmStore)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(result).To(BeTrue())
-		})
-
-		It("Should return IfNotPresent if configmap doesn't have imagePullPolicy set", func() {
-			cfgMap := kubev1.ConfigMap{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "kube-system",
-					Name:      "kubevirt-config",
-				},
-				Data: map[string]string{},
-			}
-			cmListWatch = MakeFakeConfigMapWatcher([]kubev1.ConfigMap{cfgMap})
-			cmInformer = cache.NewSharedIndexInformer(cmListWatch, &v1.VirtualMachineInstance{}, time.Second, cache.Indexers{cache.NamespaceIndex: cache.MetaNamespaceIndexFunc})
-			cmStore = cmInformer.GetStore()
-			go cmInformer.Run(stopChan)
-			cache.WaitForCacheSync(stopChan, cmInformer.HasSynced)
-
-			result, err := GetImagePullPolicy(cmStore)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(result).To(Equal(kubev1.PullIfNotPresent))
-		})
-
-		It("Should return Always if imagePullPolicy = Always", func() {
-			cfgMap := kubev1.ConfigMap{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "kube-system",
-					Name:      "kubevirt-config",
-				},
-				Data: map[string]string{ImagePullPolicyKey: "Always"},
-			}
-			cmListWatch = MakeFakeConfigMapWatcher([]kubev1.ConfigMap{cfgMap})
-			cmInformer = cache.NewSharedIndexInformer(cmListWatch, &v1.VirtualMachineInstance{}, time.Second, cache.Indexers{cache.NamespaceIndex: cache.MetaNamespaceIndexFunc})
-			cmStore = cmInformer.GetStore()
-			go cmInformer.Run(stopChan)
-			cache.WaitForCacheSync(stopChan, cmInformer.HasSynced)
-
-			result, err := GetImagePullPolicy(cmStore)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(result).To(Equal(kubev1.PullAlways))
-		})
-
-		It("Should return an error if imagePullPolicy is not valid", func() {
-			cfgMap := kubev1.ConfigMap{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "kube-system",
-					Name:      "kubevirt-config",
-				},
-				Data: map[string]string{ImagePullPolicyKey: "IHaveNoStrongFeelingsOneWayOrTheOther"},
-			}
-			cmListWatch = MakeFakeConfigMapWatcher([]kubev1.ConfigMap{cfgMap})
-			cmInformer = cache.NewSharedIndexInformer(cmListWatch, &v1.VirtualMachineInstance{}, time.Second, cache.Indexers{cache.NamespaceIndex: cache.MetaNamespaceIndexFunc})
-			cmStore = cmInformer.GetStore()
-			go cmInformer.Run(stopChan)
-			cache.WaitForCacheSync(stopChan, cmInformer.HasSynced)
-
-			_, err := GetImagePullPolicy(cmStore)
-			Expect(err).To(HaveOccurred())
-		})
-	})
 })
-
-func MakeFakeConfigMapWatcher(configMaps []kubev1.ConfigMap) *cache.ListWatch {
-	cmListWatch := &cache.ListWatch{
-		ListFunc: func(options metav1.ListOptions) (runtime.Object, error) {
-			return &kubev1.ConfigMapList{Items: configMaps}, nil
-		},
-		WatchFunc: func(options metav1.ListOptions) (watch.Interface, error) {
-			fakeWatch := watch.NewFake()
-			for _, cfgMap := range configMaps {
-				fakeWatch.Add(&cfgMap)
-			}
-			return watch.NewFake(), nil
-		},
-	}
-	return cmListWatch
-}
 
 func True() *bool {
 	b := true
