+++ pkg/virt-controller/watch/vmi_test.go
@@ -42,6 +42,7 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 
 	"kubevirt.io/kubevirt/pkg/api/v1"
+	"kubevirt.io/kubevirt/pkg/config"
 	"kubevirt.io/kubevirt/pkg/kubecli"
 	"kubevirt.io/kubevirt/pkg/log"
 	"kubevirt.io/kubevirt/pkg/testutils"
@@ -155,12 +156,12 @@ var _ = Describe("VirtualMachineInstance watcher", func() {
 
 		configMapInformer, _ = testutils.NewFakeInformerFor(&k8sv1.Pod{})
 		controller = NewVMIController(
-			services.NewTemplateService("a", "b", "c", configMapInformer.GetStore()),
+			services.NewTemplateService("a", "b", "c", config.NewClusterConfig(configMapInformer.GetStore())),
 			vmiInformer,
 			podInformer,
 			recorder,
 			virtClient,
-			configMapInformer,
+			config.NewClusterConfig(configMapInformer.GetStore()),
 			dataVolumeInformer)
 		// Wrap our workqueue to have a way to detect when we are done processing updates
 		mockQueue = testutils.NewMockWorkQueue(controller.Queue)
