+++ pkg/virt-launcher/virtwrap/manager.go
@@ -34,8 +34,8 @@ import (
 	k8sv1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"kubevirt.io/kubevirt/pkg/api/v1"
-	cloudinit "kubevirt.io/kubevirt/pkg/cloud-init"
-	"kubevirt.io/kubevirt/pkg/config"
+	"kubevirt.io/kubevirt/pkg/cloud-init"
+	"kubevirt.io/kubevirt/pkg/config-disk"
 	"kubevirt.io/kubevirt/pkg/emptydisk"
 	"kubevirt.io/kubevirt/pkg/ephemeral-disk"
 	"kubevirt.io/kubevirt/pkg/hooks"
@@ -121,11 +121,11 @@ func (l *LibvirtDomainManager) preStartHook(vmi *v1.VirtualMachineInstance, doma
 		return domain, fmt.Errorf("creating empty disks failed: %v", err)
 	}
 	// create ConfigMap disks if they exists
-	if err := config.CreateConfigMapDisks(vmi); err != nil {
+	if err := config_disk.CreateConfigMapDisks(vmi); err != nil {
 		return domain, fmt.Errorf("creating config map disks failed: %v", err)
 	}
 	// create Secret disks if they exists
-	if err := config.CreateSecretDisks(vmi); err != nil {
+	if err := config_disk.CreateSecretDisks(vmi); err != nil {
 		return domain, fmt.Errorf("creating secret disks failed: %v", err)
 	}
